combined createOrderContext and editOrderContext into one orderContext, removed unnecessary types, fixed a bug with CurrencyInput
