+++ src/index.tsx
@@ -3,7 +3,7 @@ import { HashRouter } from "react-router-dom";
 import { Provider } from "react-redux";
 
 import App from "./App";
-import { CreateOrderFormContextProvider, EditOrderFormContextProvider, MuiThemeProvider } from "./context";
+import { OrderFormContextProvider, MuiThemeProvider } from "./context";
 import store from "./store";
 
 import "./index.scss";
@@ -12,13 +12,11 @@ const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement)
 root.render(
 	<HashRouter>
 		<Provider store={store}>
-			<CreateOrderFormContextProvider>
-				<EditOrderFormContextProvider>
-					<MuiThemeProvider>
-						<App />
-					</MuiThemeProvider>
-				</EditOrderFormContextProvider>
-			</CreateOrderFormContextProvider>
+			<OrderFormContextProvider>
+				<MuiThemeProvider>
+					<App />
+				</MuiThemeProvider>
+			</OrderFormContextProvider>
 		</Provider>
 	</HashRouter>
 );
